Refactor netlink: rule import (+ new file for rules).
