+++ babeld.c
@@ -52,6 +52,7 @@ THE SOFTWARE.
 #include "resend.h"
 #include "configuration.h"
 #include "local.h"
+#include "rule.h"
 
 struct timeval now;
 
@@ -123,6 +124,22 @@ kernel_link_notify(struct kernel_link *link, void *closure)
         }
     }
     return 0;
+}
+
+static int
+kernel_rule_notify(struct kernel_rule *rule, void *closure)
+{
+    int i;
+    if(martian_prefix(rule->src, rule->src_plen))
+        return 0;
+
+    i = rule->priority - src_table_prio;
+
+    if(i < 0 || SRC_TABLE_NUM <= i)
+        return 0;
+
+    kernel_rules_changed = 1;
+    return -1;
 }
 
 int
@@ -517,6 +534,9 @@ main(int argc, char **argv)
     rc = check_xroutes(0);
     if(rc < 0)
         fprintf(stderr, "Warning: couldn't check exported routes.\n");
+    rc = check_rules();
+    if(rc < 0)
+        fprintf(stderr, "Warning: couldn't check rules.\n");
 
     kernel_routes_changed = 0;
     kernel_rules_changed = 0;
@@ -619,6 +639,7 @@ main(int argc, char **argv)
             filter.route = kernel_route_notify;
             filter.addr = kernel_addr_notify;
             filter.link = kernel_link_notify;
+            filter.rule = kernel_rule_notify;
             kernel_callback(&filter);
         }
 
@@ -690,6 +711,9 @@ main(int argc, char **argv)
             rc = check_xroutes(1);
             if(rc < 0)
                 fprintf(stderr, "Warning: couldn't check exported routes.\n");
+            rc = check_rules();
+            if(rc < 0)
+                fprintf(stderr, "Warning: couldn't check rules.\n");
             kernel_routes_changed = kernel_rules_changed =
                 kernel_addr_changed = 0;
             if(kernel_socket >= 0)
@@ -787,6 +811,7 @@ main(int argc, char **argv)
         gettime(&now);
         interface_up(ifp, 0);
     }
+    release_tables();
     kernel_setup_socket(0);
     kernel_setup(0);
